# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact: {}
  description: REST API หลังบ้านเว็บไซต์โรงพยาบาลสะเมิง (ข่าว กิจกรรม บุคลากร ITA/MOIT
    ผู้ใช้งาน สิทธิ์เมนู)
  title: Hospital Samoeng API
  version: "1.0"
paths:
...
  /hero:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: list รูปสไลด์ hero หน้าแรก (เรียงตามลำดับ)
      tags:
      - hero
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: รูปภาพ .jpg/.jpeg/.png
        in: formData
        name: image
        required: true
        type: file
      - description: ลำดับการแสดง (ยิ่งน้อยยิ่งขึ้นก่อน)
        in: formData
        name: order
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: เพิ่มรูปสไลด์ hero (admin)
      tags:
      - hero
  /hero/{id}:
    delete:
      parameters:
      - description: hero slide id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: เหลือรูปเดียว ลบไม่ได้
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ลบรูปสไลด์ hero (admin) — ห้ามลบจนเหลือ 0 รูป
      tags:
      - hero
    put:
      consumes:
      - application/json
      parameters:
      - description: hero slide id
        in: path
        name: id
        required: true
        type: integer
      - description: ลำดับใหม่ เช่น {\
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: แก้ลำดับรูปสไลด์ hero (admin)
      tags:
      - hero
  /ita:
    get:
      parameters:
      - description: filter ตาม year id
        in: query
        name: year_id
        type: integer
      - description: หน้า (default 1)
        in: query
        name: page
        type: integer
      - description: ต่อหน้า (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: data + meta{total,page,limit}
          schema:
            additionalProperties: true
            type: object
      summary: list ไฟล์ ITA (filter ตามปี + pagination)
      tags:
      - ita
  /ita/{id}:
    delete:
      parameters:
      - description: ITA file id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ไม่พบไฟล์
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ลบไฟล์ ITA (เฉพาะ admin)
      tags:
      - ita
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: ITA id
        in: path
        name: id
        required: true
        type: integer
      - description: ชื่อเอกสารใหม่
        in: formData
        name: title
        type: string
      - description: ไฟล์ PDF ใหม่ (ไม่แนบ = ใช้ไฟล์เดิม)
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: แก้ไขเอกสาร ITA — เปลี่ยนชื่อ และ/หรือ เปลี่ยนไฟล์ PDF (เฉพาะ admin)
      tags:
      - ita
  /ita/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: MOIT item id
        in: formData
        name: item_id
        required: true
        type: integer
      - description: ITA year id (ต้องตรงกับ item)
        in: formData
        name: year_id
        required: true
        type: integer
      - description: ชื่อเอกสาร (ไม่ใส่ = ใช้ชื่อไฟล์)
        in: formData
        name: title
        type: string
      - description: ไฟล์ PDF
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: อัปโหลดไฟล์ PDF เข้า MOIT item (เฉพาะ admin)
      tags:
      - ita
  /ita/years:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: list ปี ITA ทั้งหมด (เรียงใหม่→เก่า)
      tags:
      - ita
    post:
      consumes:
      - application/json
      parameters:
      - description: ปี พ.ศ. 2500-2600
        in: body
        name: body
        required: true
        schema:
          properties:
            year:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ปีไม่ถูกต้อง/ซ้ำ
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: สร้างปี ITA ใหม่ (เฉพาะ admin)
      tags:
      - ita
...
securityDefinitions:
  BearerAuth:
    description: ใส่ค่าเป็น "Bearer <token>" (ได้ token จาก POST /login)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"